+++ internal/cfg/setters.go
@@ -1,0 +1,7 @@
+package cfg
+
+func SetLoginGate(val bool) {
+	lock.Lock()
+	defer lock.Unlock()
+	globalConfig.loginGate = val
+}
